notes about javafx, and fix userDao.getByUsername
all stuff is currently on main thread so no need for runLater now.
but will once fetching data in background will need to do that.

case issue between id and username.

Exception on change user resolved by ignoring when index is -1.
note: action is triggered when choice box values are cleared and
updated.